requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
openpyxl
